Address Claude's 1st feedback round: move UserMembershipDetail type to shared user.ts types file
